requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
discogs-api==0.1.0
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.1
Pillow==8.3.2
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
wikipedia==1.4.0
